desktop-runner/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "desktop-runner"
version = "0.1.0"
description = "Desktop runner skeleton"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pywinauto>=0.6.8",
]

[tool.setuptools.packages.find]
where = ["src"]
